@@ -1,29 +1,44 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.13'
+# gem 'rails', '3.2.13'
+gem 'rails', '~>4.1.0'
+
+gem 'rails4_upgrade'
+
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
 
 # Gems used only for assets and not required
 # in production environments by default.
-group :assets do
-  gem 'sass-rails'
-  gem 'coffee-rails', '~> 3.2.1'
 
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
+# gem 'sass-rails'
+gem 'sass-rails', '~>4.0.0'
 
-  gem 'uglifier', '>= 1.0.3'
-  gem 'twitter-bootstrap-rails', '~> 2.2.8'
-  gem 'therubyracer'
-  gem 'less-rails'
-end
+# gem 'coffee-rails'
+gem 'coffee-rails', '~>4.0.0'
+gem 'sprockets-rails', :require => 'sprockets/railtie'
+
+
+# See https://github.com/sstephenson/execjs#readme for more supported runtimes
+# gem 'therubyracer', :platforms => :ruby
+
+gem 'uglifier', '>= 1.0.3'
+# gem "twitter-bootstrap-rails"
+gem 'therubyracer'
+gem 'less-rails'
+# gem 'activeresource'
+
+gem "bson"
+gem "moped", github: "mongoid/moped"
 
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
-gem 'mongoid', '~> 3.1.6'
-gem 'devise', '3.2.4'
+gem 'mongoid', '~> 4.0.0'
+# gem 'activemodel', '~> 4.1.6'
+# gem 'mongoid', '~> 4.0.0'
+# gem 'devise', '3.2.4'
+gem 'devise', '3.0.0.rc'
 gem 'omniauth', '1.2.1'
 gem 'omniauth-linkedin', '0.1.0'
 gem 'linkedin', '0.4.7'
@@ -31,17 +46,38 @@ gem 'linkedin', '0.4.7'
 gem "mongoid-paperclip", "~> 0.0.8", :require => "mongoid_paperclip"
 gem 'aws-sdk', '~> 1.3.4', :require => "aws-sdk"
 
-gem 'best_in_place', '~> 2.1.0'
-gem 'bootstrap-datepicker-rails', '~> 1.3.0.2'
-gem 'bootstrap-file-input', '~> 0.0.1'
-gem 'bootflat-rails'
-gem 'amistad', '~> 0.9.2'
+# gem 'best_in_place', '~> 2.1.0'
+gem 'best_in_place', github: "bernat/best_in_place"
+# gem 'anjlab-bootstrap-rails', :require => 'bootstrap-rails',
+#                           :github => 'anjlab/bootstrap-rails',
+#                           :branch => '3.0.0'
+# gem 'bootstrap-sass', github: 'thomas-mcdonald/bootstrap-sass'
+# gem 'bootstrap-datepicker-rails', '~> 1.3.0.2'
+# gem 'bootstrap-sass', '3.0.2.1'
+# gem 'bootstrap-file-input', '~> 0.0.1'
+# gem 'bootflat-rails'
+# gem 'amistad', '~> 0.9.2'
 gem 'faker', '~> 1.3.0'
+
+gem "twitter-bootstrap-rails"
+
+gem 'actionpack-action_caching', '~>1.0.0'
+gem 'actionpack-page_caching', '~>1.0.0'
+gem 'actionpack-xml_parser', '~>1.0.0'
+gem 'actionview-encoded_mail_to', '~>1.0.4'
+# gem 'activerecord-session_store', '~>0.0.1'
+gem 'activeresource', '~>4.0.0.beta1'
+gem 'protected_attributes', '~>1.0.1'
+gem 'rails-observers', '~>0.1.1'
+gem 'rails-perftest', '~>0.0.2'
+
 
 group :development do
   gem 'better_errors',     '1.0.1'
   gem 'binding_of_caller', '0.7.2'
+  gem 'rspec-rails', '>= 2.13.2'
 end
+
 
 # To use ActiveModel has_secure_password
 # gem 'bcrypt-ruby', '~> 3.0.0'
